fix: added import alias for providers and moved body provider there
